fix unit test for coverage
