+++ android/anc/src/main/java/org/smartregister/fhircore/anc/ui/report/ReportScreens.kt
@@ -68,6 +68,7 @@ const val TOOLBAR_BACK_ARROW = "toolbarBackArrow"
 const val REPORT_HOME_PRELOAD = "reportHomePreLoad"
 const val REPORT_MEASURE_LIST = "reportMeasureList"
 const val REPORT_MEASURE_ITEM = "reportMeasureItem"
+const val REPORT_FILTER_PAGE = "reportFiltertPage"
 const val REPORT_DATE_RANGE_SELECTION = "reportDateRangeSelection"
 const val REPORT_DATE_SELECT_ITEM = "reportDateSelectItem"
 const val REPORT_PATIENT_SELECTION = "reportPatientSelection"
