+++ android/anc/src/test/java/org/smartregister/fhircore/anc/ui/report/ReportScreensTest.kt
@@ -16,6 +16,7 @@
 
 package org.smartregister.fhircore.anc.ui.report
 
+import androidx.compose.ui.test.assertHasClickAction
 import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onNodeWithTag
 import io.mockk.spyk
@@ -37,6 +38,7 @@ class ReportScreensTest : RobolectricTest() {
     spyk(
       object {
         // Imitate click action by doing nothing
+        fun onBackPress() {}
         fun onReportMeasureItemClick() {}
         fun onStartDatePress() {}
         fun onEndDatePress() {}
@@ -64,6 +66,15 @@ class ReportScreensTest : RobolectricTest() {
   fun testReportMeasureItem() {
     composeRule.setContent { ReportRow(reportItem = ReportItem("test")) }
     composeRule.onNodeWithTag(REPORT_MEASURE_ITEM).assertExists()
+  }
+
+  @Test
+  fun testTopBarBox() {
+    composeRule.setContent {
+      TopBarBox(topBarTitle = "test", onBackPress = { listenerObjectSpy.onBackPress() })
+    }
+    composeRule.onNodeWithTag(TOOLBAR_TITLE).assertExists()
+    composeRule.onNodeWithTag(TOOLBAR_BACK_ARROW).assertHasClickAction()
   }
 
   @Test
@@ -136,18 +147,5 @@ class ReportScreensTest : RobolectricTest() {
     composeRule.onNodeWithTag(REPORT_CHANGE_PATIENT).assertExists()
     // composeRule.onNodeWithTag(REPORT_CHANGE_PATIENT).performClick()
     // verify { listenerObjectSpy.onPatientChangeClick() }
-  }
-
-  @Test
-  fun testGenerateReportButton() {
-    composeRule.setContent {
-      GenerateReportButton(
-        generateReportEnabled = true,
-        onGenerateReportClicked = { listenerObjectSpy.onGenerateReportClick() }
-      )
-    }
-    composeRule.onNodeWithTag(REPORT_GENERATE_BUTTON).assertExists()
-    // composeRule.onNodeWithTag(REPORT_GENERATE_BUTTON).performClick()
-    // verify { listenerObjectSpy.onGenerateReportClick() }
   }
 }
